Update mapping and drawing dependencies
