+++ package.json
@@ -34,11 +34,11 @@
     "@fontsource-variable/inter": "^5.3.0",
     "@heroicons/vue": "^2.2.0",
     "@iconify-prerendered/vue-mdi": "^0.28.1737398331",
-    "@lucide/vue": "^1.38.0",
-    "@orbat-mapper/control-measures": "^0.15.1",
+    "@lucide/vue": "^1.39.0",
+    "@orbat-mapper/control-measures": "^0.16.0",
     "@orbat-mapper/convert-symbology": "^1.0.2",
-    "@orbat-mapper/tactical-draw": "^0.5.5",
-    "@orbat-mapper/tactical-draw-adapter-maplibre": "^0.3.10",
+    "@orbat-mapper/tactical-draw": "0.5.5",
+    "@orbat-mapper/tactical-draw-adapter-maplibre": "^0.3.11",
     "@orbat-mapper/tactical-map-sheet": "^0.3.5",
     "@panzoom/panzoom": "^4.6.2",
     "@popperjs/core": "^2.11.8",
@@ -101,7 +101,7 @@
     "idb": "^8.0.3",
     "immer": "^11.1.18",
     "klona": "^2.0.6",
-    "maplibre-gl": "^6.6.0",
+    "maplibre-gl": "^6.7.0",
     "markdown-it": "^14.3.1",
     "markdown-it-container": "^4.0.0",
     "mgrs": "^2.2.0",
@@ -122,7 +122,7 @@
     "unist-builder": "^4.0.0",
     "vue": "^3.5.42",
     "vue-global-events": "^3.0.1",
-    "vue-router": "^5.3.0",
+    "vue-router": "^5.3.1",
     "xast-util-to-xml": "^4.0.0",
     "xastscript": "^4.0.0",
     "xlsx": "https://cdn.sheetjs.com/xlsx-0.20.3/xlsx-0.20.3.tgz"
