+++ pnpm-workspace.yaml
@@ -1,3 +1,2 @@
 overrides:
-  "@orbat-mapper/tactical-draw": "0.5.4"
   esbuild@<=0.24.2: ">=0.25.0"
